requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-genai
qdrant-client
requests